Broke pubsub module into a package and two modules
