+++ examples/websockets_lib/schema.py
@@ -7,7 +7,6 @@ from graphql_ws.pubsub import AsyncioPubsub
 p = AsyncioPubsub()
 
 # TODO:
-# - Breakup pubsub module into package and two modules
 # - Add explanation on how to use pubsub
 # - Modify author code to peform proper cleanup during cancel()
